Introduce Token Factory
the goal of the token factory is to be able to upgrade the token code
without losing any history and balances.
To do this a Token can be "forked" to a child token using a token
factory. The token factory can be updated by the owner of the token
contract. The child token is referenced from the parent token to
preserve the history.
The change of the tokenfactory will require some voting (future).